+++ web/src/__tests__/async/observation-api.servertest.ts
@@ -1,5 +1,5 @@
 import { createObservation as createObservationObject } from "@/src/__tests__/fixtures/tracing-factory";
-import { createObservations as createObservationsInClickhouse } from "@/src/__tests__/server/repositories/clickhouse-helpers";
+import { createObservations as createObservationsInClickhouse } from "@/src/__tests__/async/repositories/clickhouse-helpers";
 import { makeZodVerifiedAPICall } from "@/src/__tests__/test-utils";
 import { GetObservationV1Response } from "@/src/features/public-api/types/observations";
 import { v4 } from "uuid";
