+++ web/src/__tests__/async/traces-api.servertest.ts
@@ -5,7 +5,7 @@ import {
 import {
   createObservations,
   createTraces,
-} from "@/src/__tests__/server/repositories/clickhouse-helpers";
+} from "@/src/__tests__/async/repositories/clickhouse-helpers";
 import { makeZodVerifiedAPICall } from "@/src/__tests__/test-utils";
 import {
   GetTracesV1Response,
